Fix bug difference of time_t size between 32bit and 64bit
- Refer to :
  The IntPtr type is designed to be an integer whose size is
  platform-specific. That is, an instance of this type is expected to be
  32-bits on 32-bit hardware and operating systems, and 64-bits on 64-bit
  hardware and operating systems.

+++ src/Tizen.Applications.Common/Interop/Interop.ApplicationManager.cs
@@ -313,7 +313,7 @@ internal static partial class Interop
             internal IntPtr pkgName;
             internal IntPtr appPath;
             internal IntPtr arg;
-            internal long launchTime;
+            internal IntPtr launchTime;
             internal IntPtr instanceId;
             internal IntPtr instanceName;
             internal IntPtr icon;
